Changes based on PR Feedback
