+++ locale/de/messages.json
@@ -561,8 +561,6 @@
   "mnk.demolish.checklist.requirement.name": "Gesamtanwendungszeit von <0/>",
   "mnk.demolish.suggestion.content": "",
   "mnk.demolish.suggestion.why": "",
-  "mnk.dragonkick.suggestions.overkick.content": "",
-  "mnk.dragonkick.suggestions.overkick.why": "",
   "mnk.forms.suggestions.dropped.content": "Vermeide es, Formen zu verlieren. Du musst eventuell z.B. Rakshasas Umarmung benutzen oder einfach näher am Gegner bleiben, damit deine Kombo nicht ausläuft. Dies deutet normalerweise auf größere Probleme hin.",
   "mnk.forms.suggestions.dropped.why": "Form wurde {0, plural, one {# -mal} other {#-mal}} verloren.",
   "mnk.forms.suggestions.formless.content": "Vermeide es, Kombo-Starter außerhalb von <0/> zu benutzen, da der Form-Bonus nur in der korrekten Form aktiviert wird.",
@@ -588,6 +586,8 @@
   "mnk.steppies.suggestions.bootshine.why": "{0}-mal <0/> an der falschen Position angesetzt.",
   "mnk.steppies.suggestions.dragon_kick.content": "Verhindere <0/> ohne <1/>-Buff einzusetzen.",
   "mnk.steppies.suggestions.dragon_kick.why": "{0} Angriffspotenz durch {1}-mal nicht-aktivierten <0/>-Buff verloren.",
+  "mnk.steppies.suggestions.overkick.content": "",
+  "mnk.steppies.suggestions.overkick.why": "",
   "mnk.twinsnakes.checklist.description": "",
   "mnk.twinsnakes.checklist.name": "Halte <0/> aufrecht",
   "mnk.twinsnakes.checklist.requirement.name": "<0/> Gesamtanwendungszeit",
