+++ locale/en/messages.json
@@ -561,8 +561,6 @@
   "mnk.demolish.checklist.requirement.name": "<0/> uptime",
   "mnk.demolish.suggestion.content": "Avoid refreshing <0/> significantly before its expiration. Unnecessary refreshes risk overwriting buff snapshots.",
   "mnk.demolish.suggestion.why": "An average of {0} seconds of <0/> clipped per minute due to early refreshes.",
-  "mnk.dragonkick.suggestions.overkick.content": "Avoid using <0/> while under the effect of <1/>. Use <2/> instead.",
-  "mnk.dragonkick.suggestions.overkick.why": "You used <0/>  {0, plural, one {# time} other {# times}} while already having the <1/> buff.",
   "mnk.forms.suggestions.dropped.content": "Avoid dropping Forms. You may need to use a gap closer or stay closer to the enemy to avoid your combo timing out. This usually indicates a bigger problem.",
   "mnk.forms.suggestions.dropped.why": "Form was dropped {0, plural, one {# time.} other {# times.}}",
   "mnk.forms.suggestions.formless.content": "Avoid using combo starters outside of <0/> as the Form bonus is only activated in the correct form.",
@@ -588,6 +586,8 @@
   "mnk.steppies.suggestions.bootshine.why": "{0, plural, one {# use of} other {# uses of}} <0/> executed with incorrect position.",
   "mnk.steppies.suggestions.dragon_kick.content": "Avoid unbuffed <0/> by using <1/> before it.",
   "mnk.steppies.suggestions.dragon_kick.why": "{0} potency lost to missing <0/> buff {1} times.",
+  "mnk.steppies.suggestions.overkick.content": "Avoid using <0/> while under the effect of <1/>. Use <2/> instead.",
+  "mnk.steppies.suggestions.overkick.why": "You used <0/>  {0, plural, one {# time} other {# times}} while already having the <1/> buff.",
   "mnk.twinsnakes.checklist.description": "Disciplined Fist is an easy 15% buff to your DPS.",
   "mnk.twinsnakes.checklist.name": "Keep <0/> up",
   "mnk.twinsnakes.checklist.requirement.name": "<0/> uptime",
