+++ locale/fr/messages.json
@@ -561,8 +561,6 @@
   "mnk.demolish.checklist.requirement.name": "Temps d'activité total de <0/>",
   "mnk.demolish.suggestion.content": "Evitez de rafraichir <0/> trop de temps avec son expiration. Des applications inutiles risquent de remplacer des applications sous buffs.",
   "mnk.demolish.suggestion.why": "",
-  "mnk.dragonkick.suggestions.overkick.content": "",
-  "mnk.dragonkick.suggestions.overkick.why": "",
   "mnk.forms.suggestions.dropped.content": "Évitez de laisser tomber les formes. Vous devrez peut-être utiliser un espace plus proche ou rester plus près de l'ennemi pour éviter que votre combo ne s'arrête. Cela indique généralement un problème plus important.",
   "mnk.forms.suggestions.dropped.why": "Position perdu {0, plural,one {fois.}other {#fois.}}",
   "mnk.forms.suggestions.formless.content": "",
@@ -588,6 +586,8 @@
   "mnk.steppies.suggestions.bootshine.why": "{0, plural,one {#utilisation de }other {#utilisations de}}<0/> exécuter avec des positions incorrect.",
   "mnk.steppies.suggestions.dragon_kick.content": "Évitez les <0/> non buffée en utilisant <1/> avant.",
   "mnk.steppies.suggestions.dragon_kick.why": "{0} puissance perdue en raison de l'absence de <0/> buff {1} fois.",
+  "mnk.steppies.suggestions.overkick.content": "",
+  "mnk.steppies.suggestions.overkick.why": "",
   "mnk.twinsnakes.checklist.description": "Poings vigoureux est un buff facile de 15% à votre DPS.",
   "mnk.twinsnakes.checklist.name": "Gardez <0/> actif",
   "mnk.twinsnakes.checklist.requirement.name": "<0/> temps d'activité",
