+++ locale/ja/messages.json
@@ -561,8 +561,6 @@
   "mnk.demolish.checklist.requirement.name": "アクティブ率: <0/>",
   "mnk.demolish.suggestion.content": "効果時間が残っている<0/>を上書きしないようにしましょう。不必要に更新すると、デバフの威力を損うおそれがあります。",
   "mnk.demolish.suggestion.why": "",
-  "mnk.dragonkick.suggestions.overkick.content": "",
-  "mnk.dragonkick.suggestions.overkick.why": "",
   "mnk.forms.suggestions.dropped.content": "型の効果が落ちないように頑張ってください。GCDやコンボが続けるめに、できれば敵から離れないようにしてください。そうすると型の効果が更新されます。",
   "mnk.forms.suggestions.dropped.why": "型の効果時間が切れた回数：{0, plural, one {#回} other {#回}}",
   "mnk.forms.suggestions.formless.content": "",
@@ -588,6 +586,8 @@
   "mnk.steppies.suggestions.bootshine.why": "<0/>は対象の背中から{0, plural, other {#回}}実行できませんでした。",
   "mnk.steppies.suggestions.dragon_kick.content": "バフされていない<0/>を遠慮してください。先に必ず<1/>を実行してください。",
   "mnk.steppies.suggestions.dragon_kick.why": "<0/>のバフが {1} 回追加していないため、{0} 威力がなくなりました。",
+  "mnk.steppies.suggestions.overkick.content": "",
+  "mnk.steppies.suggestions.overkick.why": "",
   "mnk.twinsnakes.checklist.description": "",
   "mnk.twinsnakes.checklist.name": "<0/> を維持しましょう",
   "mnk.twinsnakes.checklist.requirement.name": "アクティブ率: <0/>",
