+++ locale/ko/messages.json
@@ -561,8 +561,6 @@
   "mnk.demolish.checklist.requirement.name": "<0/> 업타임",
   "mnk.demolish.suggestion.content": "이전에 부여한 <0/> 지속 시간이 충분하다면 갱신하지 마십시오. 불필요하게 갱신하면 버프 위력을 잃을 수 있습니다.",
   "mnk.demolish.suggestion.why": "",
-  "mnk.dragonkick.suggestions.overkick.content": "",
-  "mnk.dragonkick.suggestions.overkick.why": "",
   "mnk.forms.suggestions.dropped.content": "품새를 항상 유지하십시오. 콤보 시간이 만료되지 않도록 적에게 가까이 붙거나 돌진기를 사용하십시오. 이 항목이 표시되었다면 더욱 심각한 문제가 있을 가능성이 높습니다.",
   "mnk.forms.suggestions.dropped.why": "품새를 {0, plural, other {#회}} 잃었습니다.",
   "mnk.forms.suggestions.formless.content": "<0/> 외의 구간에서 원숭이 품새 기술을 사용하지 마십시오. 품새 보너스는 오직 알맞은 품새 상태에서만 받을 수 있습니다.",
@@ -588,6 +586,8 @@
   "mnk.steppies.suggestions.bootshine.why": "<0/> 기술을 잘못된 방향에서 {0, plural, other {#회}} 사용했습니다.",
   "mnk.steppies.suggestions.dragon_kick.content": "<1/> 기술을 미리 사용하여 버프 없는 <0/> 사용을 피하십시오.",
   "mnk.steppies.suggestions.dragon_kick.why": "<0/> 버프를 {1}회 놓쳐 위력을 {0} 잃었습니다.",
+  "mnk.steppies.suggestions.overkick.content": "",
+  "mnk.steppies.suggestions.overkick.why": "",
   "mnk.twinsnakes.checklist.description": "공력은 피해량 15%를 손쉽게 증가시키는 버프입니다.",
   "mnk.twinsnakes.checklist.name": "항상 <0/> 유지하기",
   "mnk.twinsnakes.checklist.requirement.name": "<0/> 업타임",
