+++ locale/zh/messages.json
@@ -561,8 +561,6 @@
   "mnk.demolish.checklist.requirement.name": "<0/> 覆盖率",
   "mnk.demolish.suggestion.content": "避免提前刷新<0/>。不必要的刷新会覆盖掉团辅快照。",
   "mnk.demolish.suggestion.why": "因提前刷新导致每分钟平均损失了{0}秒的<0/>",
-  "mnk.dragonkick.suggestions.overkick.content": "",
-  "mnk.dragonkick.suggestions.overkick.why": "",
   "mnk.forms.suggestions.dropped.content": "避免失去身形。你可能需要突进技或保持更小的距离来避免你的连击因超时而中断。这通常意味着更大的问题",
   "mnk.forms.suggestions.dropped.why": "丢失了{0, plural,one {#次}other {次}}身形。",
   "mnk.forms.suggestions.formless.content": "避免使用除<0/>作为你的连击起手，只有当你进行正确的连击时才会获得连击的增益。",
@@ -588,6 +586,8 @@
   "mnk.steppies.suggestions.bootshine.why": "{0, plural, one {# 次用} other {# 次用}} <0/> 打中错误的身位.",
   "mnk.steppies.suggestions.dragon_kick.content": "避免在使用 <1/> 之前使用 <0/> 。",
   "mnk.steppies.suggestions.dragon_kick.why": "{1} 次因为没有 <0/> Buff，损失了 {0} 的技能威力。",
+  "mnk.steppies.suggestions.overkick.content": "",
+  "mnk.steppies.suggestions.overkick.why": "",
   "mnk.twinsnakes.checklist.description": "功力能让自己的所有伤害直接提高15%。",
   "mnk.twinsnakes.checklist.name": "保持 <0/> 不断",
   "mnk.twinsnakes.checklist.requirement.name": "<0/> 覆盖率",
